Rename C++ headers to hpp and update includes
