+++ src/non_copyable.hpp
@@ -17,3 +17,4 @@ struct NonCopyable {
     NonCopyable(const NonCopyable&) = delete;
     NonCopyable& operator=(const NonCopyable&) = delete;
 };
+
