+++ src/server/protocol.hpp
@@ -4,9 +4,6 @@
 namespace server {
 
     static const char MSG_SEPARATOR = 0x1F;
-
-    // TODO: add error generator inline func ('ERROR: <reason>') + convert this to .hpp
-
     static constexpr const char* OK = "OK";
     static constexpr const char* NOTHING = "(nil)";
     static constexpr const char* KEY_NOT_EXISTS = "ERROR: Key does not exist";
