+++ src/server/server.cpp
@@ -1,4 +1,4 @@
-#include "server.h"
+#include "server.hpp"
 #include <unordered_map>
 
 using namespace server;
