+++ src/server/shard.cpp
@@ -1,4 +1,4 @@
-#include "shard.h"
+#include "shard.hpp"
 
 using namespace server;
