+++ src/server/sockutils.cpp
@@ -1,4 +1,4 @@
-#include "sockutils.h"
+#include "sockutils.hpp"
 
 int setNonBlocking(int fd) noexcept {
   int flags = fcntl(fd, F_GETFL, 0);
